Add provided deps query support to android_library
Summary:
For global code generation, it's beneficial to use provided_deps instead of deps
so that we do not accidentally package items into the app that are used by metagen. This change
allows us to use provided_deps instead, and it also adds better syntax for referencing the declared deps.

Test Plan: new integration test

Reviewed By: dreiss

fbshipit-source-id: a796f60