Pass around one CoercedTypeCache
Summary:
Right now, each ConstructorArgMarshaller (and there are a few) news up
their own cache.

Also, some things are passed a ConstructorArgMarshaller just to use the
cache.

Also, there's a second class which creates a cache of the same data,
which I'm going to get rid of shortly.

A strong argument could be made that this should just be a static field
somewhere.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 0a0d33b