Re-use CoercedTypeCache in TargetNodeFactory
Summary:
There is a slight change in performance characteristic here.
Before, we cached a filtered list of ParamInfos, filtered to only those
which are deps and inputs and match relevant classes. Now, we perform
that filtering after fetching from the cache.

The performance effect is negligible because very few arg params
don't match the above criteria, and the instanceof check is fast.

Reviewed By: Coneko

fbshipit-source-id: 86cd14f