aapt2: Add support for android_prebuilt_aar and react_native_library
Summary:
Both of these produce Android resources, so they must be able to supply
aapt2_compile rules to work with aapt2.

ReactNativeLibraryGraphEnhancer being in a separate package means that
Aapt2Compile and AAPT2_COMPILE_FLAVOR need to be public.

I'm open to discussion for whether this is the right approach.  Is it
reasonable to require every resource-supplying rule to support the
aapt2_compile flavor, or should android_binary create those rules for
them?

Test Plan:
CI.
Built and ran FB4A with it.

Reviewed By: asp2insp

fbshipit-source-id: 7f0b665