Aapt2 support for android_prebuilt_aar and react_native_library (2nd try)
Summary:
Both of these produce Android resources, so they must be able to supply
aapt2_compile rules to work with aapt2.

ReactNativeLibraryGraphEnhancer being in a separate package means that
Aapt2Compile and AAPT2_COMPILE_FLAVOR need to be public.

Unlike the first land attempt, in ReactNativeLibraryGraphEnhancer, we
append the aapt2 flavor to the AndroidResource rule so that we don't get
target collisions with different RN flavors.

Test Plan:
CI.
Built and ran FB4A with it.

fbshipit-source-id: c78f69d