Use optimized finder for Xcode tool paths
Summary:
This used to use ExecutableFinder, which was originally used to find
executables on the PATH. It ends up doing extra work in this case:

* It stats the name passed in, which for xcode toolchain purposes will never
  match, since we pass in a tool name that's never going to be in the present
  directory (nor would we want it to be).
* It will concat the name to every search path and stat them in turn, since we
  search SDK and platform directories first, they almost never match.

This diff introduces a simpler XcodeToolFinder, it does not stat the name
passed in, and it keeps a cache of directory contents for every search
directory, so non-matching files can be rejected quickly. Additionally, we use
a single instance of this class for lookups of all platforms, so that there's
further cache sharing.

Test Plan: CI

Reviewed By: illicitonion

fbshipit-source-id: 12962f1