Construct abiClasspath with only the deps that Java libraries will compile against
Summary:
This contains the initial changes for compiling against ABI JARs

`abiClasspath` was being constructed incorrectly by being supplied all deps, including annotation processor deps. This change makes it so that `abiClasspath` is supplied only the deps that we use to compile against (declared deps, provided deps, exported deps, and their transitive exported deps).

This incorrect `abiClasspath` construction actually introduced a bug where removing a dependency, but not changing the source didn't cause a rebuild. This is because the compilation classpath and annotation processor classpath were not separated. An example of this bug was found in a Buck file that had `//third-party/java/guava:guava` as a dep, which is also commonly a dep of annotation processors. Removal of this dep didn't actually cause a change in the rule key. This caused the build to be successful when it shouldn't have been.

Test Plan:
- Added a new test in `DefaultJavaLibraryTest` that checks the dep file predicate for annotation processor deps
- Tested locally by reproducing the issue without the changes, then adding the changes and verified a re-build occurs due to a different rule key being calculated

Reviewed By: jkeljo

fbshipit-source-id: d8b592e