parser: fix `read_config()` invalidation bug
Summary:
The parser tracks which `.buckconfig` settings a build file reads while parsing
so it knows to invalidate the parsed build file from the cache when the config
setting changes.  However, there was a bug in how "empty" config settings were
handled (e.g. `[foo] bar=`).  When pass in these config settings to the build
file parser, we use the raw config which parses empty config settings as present
empty strings, but when we verify if a cached node has been invalidated due to
config changes, we parse empty config settings as empty optionals.  This means
Buck *always* invalidates nodes which use `read_config` on empty config settings.

This diff changes the invalidation logic to use the raw config style of reading
values to make it consistent.  It's not really that important (at least from a
parser standpoint) which style we use, as long as it's consistent.

Test Plan: added unittest

Reviewed By: mzlee

fbshipit-source-id: 6502dd8