Reduce DefaultJavaLibrary creation copypasta: CalculateAbi for non-Android targets
Summary:
`CalculateAbi` is called in exactly the same way across most targets. That way takes
about six lines of code, and is about to get more complicated as we move to generating
ABI jars in different ways. This commit consolidates most usages of `CalculateAbi`
into `DefaultJavaLibraryBuilder` itself. The Android rules are more complex because of
the interaction with `DummyRDotJava`; that will get untangled in a later commit.

Test Plan: CI builds

Reviewed By: asp2insp

fbshipit-source-id: 1a4deb5