Add some controls for compiling ABIs from source
Summary:
There are a variety of steps to migrating a codebase to support
generating ABIs from source, and a variety of ways in which things
can go wrong. These controls are intended to enable people to deal
with both.

See comments on each one for what it's for.

Test Plan: Use them

Reviewed By: dreiss

fbshipit-source-id: d100c7f