Compile against ABI JARs in DefaultJavaLibrary
Summary:
The Java compiler only cares about the ABIs on classes on the classpath and the existing behaviour is that the full JARs containing the complete implementation of classes are being supplied.

With this diff, JARs containing only the ABI of classes will be passed Java compiler.  We'll reduce I/O with less code to load into memory this way and should see a small win in build times with this. This diff also provides a good base going forward for making further modifications to ABI generation without it breaking compiling against ABIs.

Changes:
- `DefaultJavaLibrary` was modified to just pass the `Path`s contained within `abiClasspath`, since it only contains the ABI JAR paths needed to compile against
-  `DefaultClassUsageFileReader` had to be updated to reflect the change of compiling against ABI jars
- `DefaultJavaLibrary` passes `deps` to `DefaultClassUsageFileReader` when creating the list of inputs after building locally. Since it was found unnecessary to include the annotation processor path to the dep file, we can just rely on the already-constructed set of `SourcePath`s in `abiClasspath` to grab the usage entries from the dep file. This is because only compilation class path based entries will exist in the dep file.

Test Plan:
- Added tests in `StubJarTest` that cover cases that may be problematic and updated existing tests to reflect changes
- Ran diffs on build outputs to ensure compiling against ABIs did not cause a change in output JARs
  - This was done by compiling against full and ABI jars, grabbing their respective buck-out folders, and running either `find . -name '*.jar'|grep -v 'abi.jar'|xargs -I x diff x ../buck-out/x` or using Beyond Compare

Reviewed By: jkeljo

fbshipit-source-id: 79d3c1d