Make it explicit that deps are non-ABI rules
Summary:
There are upcoming changes that translate deps to their ABI versions depending on if we're compiling against ABIs or not. With this, it becomes really important to still pass the non-ABI versions of deps, regardless of the ABI compilation flag being enabled or not. In addition, it could get confusing to know what type of rules we're dealing with in the code.

These changes make the variable names of deps more explicit that they're non-ABI rules in the places they should be.

Test Plan: CI

Reviewed By: jkeljo

fbshipit-source-id: c6f94f5