Framework dependency traversal should stop at bundles.
Summary:
It doesn't make sense for a bundle to traverse through its bundle dependencies
and request frameworks of its binary.

This logic also conflicts with logic in AppleBinary, which expects Framework
Dependency queries to be made with a build target having platform flavors,
which this method does not attach to the build target.

Test Plan: I should probably add a test for this.

Reviewed By: ryu2

fbshipit-source-id: 7128acd