Lookup xcodebuildversion using plist instead of executing xcodebuild
Summary:
The code asserts that this xcode build version string can only be derived from
`xcodebuild -version`, but this is not true. The version string is actually
inside version.plist.

I verified that changing the string in there changes what xcodebuild reports as
the build version, so it's indeed the same string.

Test Plan: CI

Reviewed By: ryu2

fbshipit-source-id: 224438d