One ObjectMapper is statically shared globally
Summary:
Modifies (almost) all users of ObjectMapper to use the immutable
ObjectReader and ObjectWriter instead.

A handful of places where ObjectMappers are newed up and reconfigured
and moved to use ObjectMappers.legacyCreate.

This guarantees that we re-use the same ObjectMappers, so hit internal
caches. It also stops us having to pass around ObjectMappers through the
entire stack.

I feel dirty for introducing a static shared instance, but those are
genuinely the correct semantics for this object.

Test Plan: CI. I have manually inspected every reference to ObjectMapper in the codebase.

Reviewed By: marcinkosiba

fbshipit-source-id: 5c7af42