Add tracing to action graph queries
Summary:
deps_query and query_outputs are some of the most expensive operations
we do during action graph creation.  Add trace events for them to make
them easier to see.

I opted to not include the query text.  It's generally easy enough to
find in the source code, and I want to err on the side of lower
overhead.

Test Plan: Built FB4A.  Saw traces.

Reviewed By: jkeljo

fbshipit-source-id: ea95b61