Fix buck project focused target matching when root cell has a name
Summary:
This old code adds a cell name if one is present for a given cell path, and
then match by the full unflavored target. When the root cell has a cell name,
this approach breaks since targets in the root cell are usually referred to
without cell names, and later on target objects that represent the same target
fails to match.

Since a build target's equality for all non-parsing/display purposes is
determined by the cell path and not the cell name, I change the implementation
to compare the cell path instead of the cell path AND the cell name.

Test Plan: CI, the existing test in ProjectIntegrationTest still pass.

Reviewed By: dinhviethoa

fbshipit-source-id: f9dafe7