Fix the optimized extractZipFile logic
Summary:
The directory scanning optimization was incorrect if a zip file
contains paths that are only referenced as "part" of another, e.g., if the
contents are `foo/` and `foo/bar/baz`.  This diff should fix that, by deferring
the new file creation until after the deletions are done.  In the common case
(where paths are not skipped), this will be equivalent, but if paths are
skipped, we re-create them before writing the file.

Test Plan: CI

Reviewed By: marcinkosiba

fbshipit-source-id: da28db7