Fix non-deterministic order in library's extra classpath
Summary:
The order of classes in library xml is non-deterministic. This should not affect users (unless they commit xmls), but it makes testing harder. Also, we don't have integration tests that cover this area.

This change changes the order of files to be sorted and adds a unit test.

Test Plan: Unit tests

Reviewed By: JonShemitz

fbshipit-source-id: d227b37