Add parsing+serialization of Android's arsc and compiled xml
Summary:
The complex part here is mostly just getting the formats of things
correct. The comments contain specifications of the binary
format of the various structures (including parts that aren't yet used
by BUCK).

The formats of things are derived from
https://android.googlesource.com/platform/frameworks/base/+/kitkat-release/include/androidfw/ResourceTypes.h
and
https://android.googlesource.com/platform/frameworks/base/+/kitkat-release/libs/androidfw/ResourceTypes.cpp
and a bit of manual inspection and from some other files around there
and from the aapt source.

Test Plan: Added tests

Reviewed By: dreiss

fbshipit-source-id: dda6431