Concurrent target re-compiles MostExecutors
Summary:
Rather than exporting the non-visible dep (subverting visibility
entirely), have both rules compile this class. This will not cause
duplicate class issues, because the two targets in this BUCK file are
exclusive - the testrunner one should only be used in the testrunner and
the other one should only be used in Buck.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 586d38a