Make TrimUberRDotJava's input path optional
Summary:
Current TrimUberRDotJava checks if its input path exists on disk and
changes its behavior on that. This implicitly depends on
AaptPackageResources' behavior (and it's probably possible to trigger
incorrect behavior by a sequence of builds and changing target
specifications).

Instead, make the AaptOutputInfo path optional and pass that to
TrimUberRDotJava and switch behavior on the presence/abscence of the
optional.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 3b40dd2