AndroidLibraryDescription: compute deps query lazily
Summary:
The deps query was being computed eagerly, even for `#abi` flavors where it would
be thrown away. For global rules in large graphs, this represented a nontrivial amount
of time during action graph creation. Making it lazy avoids the computation altogether
for cases where it's not needed.

Test Plan: CI for regressions. Built a target with a deps query in it, verified in traces that there was only one query being done at action graph creation time.

Reviewed By: dreiss

fbshipit-source-id: 8564164