Embed prebuilt frameworks that are direct dependencies of apple_bundles regardless of #include-frameworks
Summary:
Currently, the prebuilt framework embedding logic is intertwined with
generate-libraries-as-frameworks logic. This diff provides a hack that allows
prebuilt frameworks to be directly included as an embedded framework, without
dealing with #framework and #include-frameworks at all.

In the long term, we should create a more robust method for specifying which
libraries should be frameworks, and which frameworks should be embedded.

Test Plan:
add a prebuilt framework as a dependency to an apple test, see that
it's included in the test bundle. Ditto with an app bundle.

Reviewed By: ryu2

fbshipit-source-id: 00d82fb