Add NativeLibraryProguardGenerator
Summary:
This constructs a buildrule that creates a proguard config file that
will be used for the proguarding the android_binary.

Uses android_binary's native_library_proguard_config_generator as an
executable. Passes it an output file followed by a list of paths for all
the outputs from all the CopyNativeLibraries.

Test Plan: Added test.

Reviewed By: marcinkosiba

fbshipit-source-id: ad3b6aa