Apple Library: add option to include headers in build phase for dynamic and static frameworks
Summary:
Adds an option, enabled by default, to Xcode project generation so that headers are added to a `Headers` builds phase and correctly bundled in the resulting `.framework` product. This behaviour can be controlled globally through `buckconfig` with a new setting, as shown:

    [apple]
        enable_framework_headers = false

Test Plan: Integration tests and internal FB project tests.

Reviewed By: dinhviethoa

fbshipit-source-id: 3784ba9