Send top-level HumanReadableException to frontend.
Summary:
HumanReadableExceptions are only caught in Main.java, everything has already been closed.
This diff catches HumanReadableException (and IOException for device out-of-space errors) in
DistBuildRunCommand::runWithoutHelp so that the message can be posted back to the frontend.

Test Plan:
Ran a distributed build forcing a HumanReadableException. The message was succesfully transmitted
over to the client SuperConsole.

Reviewed By: alisdair04

fbshipit-source-id: dfd2143