Quick hack to enable javac plugin tests to run in IntelliJ
Summary:
Buck is becoming more dependent on its `javac` plugin, and integration tests
were not able to find the plugin when run from IntelliJ.

This commit is a quick hack to fix the issue; a more correct fix will come later.

The IntelliJ build doesn't have the same module divisions as the Buck and Ant builds
do -- the plugin and bootstrapper are both built together with the main Buck source.
This commit introduces an empty plugin jar as a resource in the IntelliJ build. That
enables `PluginLoader` to find the jar and create a `ClassLoader` over it. It doesn't
matter that the jar is empty, because all the classes that we want from it are
actually loaded by the parent `ClassLoader` anyway when run from IntelliJ.

The more correct fix is to make the IntelliJ build better match the Buck and Ant
builds -- that is, have separate modules for the bootstrapper and `javac` plugin.
That's going to require moving stuff around and changing those other builds as well,
which I don't have time for presently but will be able to address soon.

Test Plan:
Run `AndroidBinaryIntegrationTest` and all tests under `com/facebook/buck/jvm/java/abi`
and `com/facebook/buck/jvm/java/plugin` within IntelliJ, observe they pass.

CI for tests run in Buck

Reviewed By: nayrmi

fbshipit-source-id: 6da31e1