rust: add #check flavour for check builds
Summary:
The Rust compiler, rustc, supports a mode of operation which can
quickly check code for syntax and type correctness, which skips all
codegen. For libraries, it also generates a metadata file containing
the public API of the library which checks on the dependents can
quickly consume. (Check builds on binaries produce no output except
for warnings, if any.)

Test Plan: CI

Reviewed By: marcinkosiba

fbshipit-source-id: 37851ea