Use ImmutableMap instead of TreeMap/LinkedHashMap to save memory
Summary:
`TreeMap.Entry` and `LinkedHashMap.Entry` are each 40 bytes on 64-bit machines.
`ImmutableMapEntry` is 24 bytes. This can make a difference when working with
really big builds.

Unfortunately `ImmutableMap` doesn't allow `null` values. For now we lose some
efficiency to deal with that safely; eventually it might be nice to carry the
sentinel through to all the places these maps are used.

Test Plan: CI for regressions

fbshipit-source-id: feaa81b