Support something like javac Plugins for Buck-internal use
Summary:
`TaskListener`s get great access to events coming out of the compiler, but
it's annoyingly difficult to use them with `Elements` or `Types` interfaces in
`javac` 7, because calling `JavacTask.getElements` or `JavacTask.getTypes` in that
version causes it to ignore subsequent calls to `setTaskListener`. We have to
lazily initialize them.

To better support lazy initialization of `TaskListener`s, this commit introduces
`BuckJavacPlugin`, which is identical to the `javac` 8 `Plugin` interface but
defined within  Buck so that it can be used even when the compiler is older than
`javac` 8. We do want to support `javac` 8+ `Plugin`s in Buck soon, so having this
concept available is a step in the right direction.

Test Plan: New unit test

Reviewed By: asp2insp

fbshipit-source-id: 7985564