Don't generate directory entries in code jars
Summary:
When spooling directly to jar we're adding directory entries to the jar.
While this is also the behavior of the command-line tool, it's not strictly
necessary and it's not the behavior of the code that generates our ABIs
from class files.

Since the code to generate ABIs from source uses the same path as the direct
to jar spooling, this results in spurious differences that make it hard to
tell if we've got the ABI generation completely right.

Since it's not necessary and just wastes (a small amount) of size, this
commit removes the directory entries from code jars.

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: 7338310