AndroidLibrary.Builder: use a different field for AndroidManifest
Summary:
`manifestFile` in an `android_library` rule means `AndroidManifest.xml`,
whereas in a `java_library` it means `META-INF/MANIFEST.MF`.

To make this (a little) clearer, and to avoid problems down the line, this
commit makes `AndroidLibrary.Builder` override the inherited
implementation of `setManifestFile`.

Test Plan:
CI

(Also this fixed an issue a commit I'm working on for generating ABIs from
source where it was trying to parse an `AndroidManifest.xml` as a `MANIFEST.MF`.)

Reviewed By: asp2insp

fbshipit-source-id: f0108ea