Add integration test for headers-only caching
Summary:
Earlier this year, yiding made it possible for targets to depend just on the headers of other targets they care about. This is great, and allowed for a significant speedup of our dylib builds. Unfortunately the builds are still not where I believe they theoretically should be, so I want to investigate a bit.

To start with, I added an integration test that verifies whether changing the implementation of a `#headers` dependency causes a target to get rebuilt.

Test Plan: beholdunittests

Reviewed By: Coneko

fbshipit-source-id: 0a1aa9a