Force targets to be processed in deterministic order
Summary:
The content of `iml` files depends on the order of targets processed in `IjModuleFactory.createModule`. Since the order is arbitrary, the content may be different on every run.

The solution is to sort target nodes in `IjModuleFactory.createModule`.

Test Plan: Added a new unit test. This test is flaky with old logic, but should always pass with the new logic.

Reviewed By: JonShemitz

fbshipit-source-id: c7af1aa