Add integration test to verify resource copying
Summary:
I got passed a bug about resource copying which is inflating our IPA size. ryu2 basically already fixed this bug, but due to a change in strategy with how we set up dylibs we exposed the same bug again.

Before I fixed it, I wanted to see if I could get a barebones integration test set up that verified that what we had worked in the hopes that I could easily modify it to expose the current bug. This diff accomplishes both - it mainly adds an integration test to verify that dynamically linking against some library doesn't cause its resources to get included in your bundle. It also contains a commented out line which adds a dependency on the libraries headers. If you uncomment this line then the bug I discussed before is exposed (aka it causes the test to fail), which means that to fix the bug all I have to do is get the unit test to pass. Exactly my kind of thing. That will be done in a followup diff.

Test Plan: `AppleBundleIntegrationTest.onlyIncludesResourcesInBundlesWhichStaticallyLinkThem`

Reviewed By: ryu2

fbshipit-source-id: 6c05b4e