Don't include resources in the bundle which pass `#header` boundaries
Summary: When we switched dylib strategies our IPA size regressed a fair amount. The reason for this was duplicated resources - ryu2 had fixed the bug originally, but when we changed our strategy from depending on `DylibBinary#shared` to depending on `LibraryInsideDylibs#headers` this logic broke. This diff fixes it.

Test Plan: `ant test`

Reviewed By: ryu2

fbshipit-source-id: 8958ba8