Don't reconstruct TargetGraph in BuckQuery
Summary:
The query implementation would, for every new node, reconstruct
the TargetGraph that it was operating over. This would lead to pretty
bad behavior when long lists of targets were passed to the query
command.
This change addresses the problem by having the query incrementally
build up a mutable graph.

Test Plan: CI

Reviewed By: illicitonion

fbshipit-source-id: d18c4bb