DefaultJavaLibrary: Map deps to SourcePaths while building the rule
Summary:
Currently `DefaultJavaLibrary` waits until it's building steps to do the mapping.
It's only getting away with this because it's using something other than the actual
classpath to build the rule key (something that will hopefully change when compile against ABI
is on all the time). However, having it take in the steps from outside makes it easier
to share logic with the ABI from source rule that I'm about to introduce.

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: a41143e