BuildTargetParser canonicalize cell names used in build targets.
Summary:
Add a method in CellPathResolver to generate a root-cell-centric cell name (a
cell name that can be resolved when given to the root cell's cell name
resolver). This meas that no-cell name means "target belonging to the root cell".

Use this method to canonicalize cell names entering the system via
BuildTargetParser.

The goal here is to make a build target's fully qualified name be isomorphic to
the build target itself for a given configuration of cells. This was previously
not the case since targets in cells may refer to targets in the same cell
without a cell name. These are now normalized to always contain a cell's name
(or vice versa, if a root cell target is referred to with a cell name).

Test Plan: CI

Reviewed By: marcinkosiba

fbshipit-source-id: e7baa6a