Update prebuilt fbxctest in buck to handle new ios versions in simulator tests
Summary:
The old version we had didn't handle ios 10.3, so simulator tests were failing
because it was seeing no valid versions.

This diff updates the build of fbxctest, which not only has support for 10.3,
but also purportedly support for future minor versions too.

Test Plan: AppleTestIntegrationTests that interact with simulator are now passing

Reviewed By: jkeljo

fbshipit-source-id: d207299