Implement correct aggregation algorithm
Summary:
Aggregation of module does not always produces correct structure of modules.

For example, in the following situation:

```
//a:a (Android)
 +-> //a/b:b (Java 8)
      +-> //a/b/c:c (Android)
```

The algorithm would detect that `//a/b:b` has a non-default Java language level and merge everything under `a/b` into one module. As a result, classes in `a/b/c` use Java SDK, not Android SDK.

The solution is a new aggregation algorithm based on a prefix tree. This prefix trees emulates filesystem structure but only for modules and allows aggregating modules at a particular level based on module types.

With this algorithm `java_library` is never merged with `android_library` even if its language level is default. This is because Java SDK and Android SDK contain different set of classes.

Test Plan: Unit tests. Also, running `buck project` on a real repo and confirm that all changes are legit.

Reviewed By: asp2insp

fbshipit-source-id: b6e4625