Remove --build-with-buck.
Summary:
This is currently unused.

Removing it makes some of the other parts in the project generator
easier to reason about. If we were to want to do this again I believe
it would be better to implement it as a completely separate project
generator that didn't take the configs into account at all, and would
translate the Buck builds to Xcode project, rather than the Buck
project generation configurations and attributes.

Test Plan: CI

Reviewed By: dinhviethoa

fbshipit-source-id: 91eae2c