Use ABI BuildRules for Java/Android library BuildRuleParams
Summary:
When compiling against ABIs, the final `BuildRuleParams` in `DefaultJavaLibraryBuilder` is constructed with all its deps translated to their ABI rules (only the appropriate ones).

`BuildRule`s serve two main purposes: to be used by `CachingBuildEngine` for building and to be used for constructing other `BuildRule`s. By focusing on `BuildRuleParams` to hold the ABI dependencies, we can have `CachingBuildEngine` have the benefits of rules depending on just their ABI dependencies and still maintain the functionality of having the full dependencies for when rules are queried for them.

Test Plan:
- Added test that utilizes the dir cache and:
    1. has A depend on B, C, D
    2. builds A
    3. performs a clean
    4. re-builds A
    5. verifies that a cache hit on B#abi, C#abi, D#abi
    6. verifies that B, C, D were not part of the build log (since we don't need to depend on them for the sake of re-building A)
- CI
- Built a very large project locally

Reviewed By: dreiss

fbshipit-source-id: a2af857