Ensure entries spooled directly to jar are written in sorted order
Summary:
We were depending on the compiler to write these in a deterministic order. And it does, but
there are no guarantees about the behavior of annotation processors or other plugins.

This change makes it such that files written by the compiler are written to the jar in sorted order,
followed by any other files that were needed (also in sorted order). That's still not identical
to `spool_mode = intermediate_to_disk`, but it is at least deterministic even when the compiler or
its plugins are not.

Test Plan: `JavaInMemoryFileObjectTest`, `JavaInMemoryFileManagerTest`, and `DefaultJavaLibraryIntegrationTest`

Reviewed By: asp2insp

fbshipit-source-id: 283b5b8