Implement stub generation to JavaFileManager in a post-enter callback
Summary:
This commit introduces a more performant and complete implementation of
stub jar generation in a post-enter callback. It's more complete in that it
generates a manifest and preserves resources output by other annotation processors
(by virtue of using the `JavaFileManager` to output the stubs). It's more performant
in that it doesn't force annotation processing rounds when there wouldn't otherwise
be any.

Test Plan: Updated unit tests

Reviewed By: asp2insp

fbshipit-source-id: 1f61b77