Put (only) inner classes referenced in the ABI in InnerClasses table
Summary:
The `InnerClasses` table in a `.class` file includes any inner classes referenced
from within that `.class` file, regardless of what outer class contains them. This
is necessary so that the JVM knows how to interpret their fully qualified class name --
`$` is a legal character in class names, so is `Foo$B` referring to an inner class of
`Foo` named `B`, or a class named `Foo$B`?

Previously we were including in the stub all non-private entries in the `InnerClasses`
table from the full class. This was both too much and too little -- it included
references to inner classes from within method bodies, and it left out references to
private inner classes from the ABI. Plus, it would be extra work to generate the useless
references in ABIs from source.

This commit makes the `InnerClasses` table construction smarter, including all inner
classes that are referenced in the ABI.

Test Plan: New unit test

Reviewed By: dreiss

fbshipit-source-id: dcb42a1