Allow android_library/apple_library dependencies for `js_bundle`
Summary:
Depending on the use case, it might be feasible to pass dependencies to `js_bundle` that in turn depend on `js_library` rules.

This adds functionality to walk a subtree of dependencies and extract all `js_library` targets from it. The implementation traverses declared dependencies of `android_library` and `apple_library` rules.

Test Plan:
- new unit tests
- new integration test

Reviewed By: marcinkosiba

fbshipit-source-id: 2cc0de9