Improve error reporting when jars are corrupted
Summary:
We had a user with a corrupted jar on disk, and it was difficult to
track down due to inadequate logging.  This diff includes two
improvements.

- JarDirectoryStep was catching and logging exceptions to the console,
  then returning a StepExecutionResult.ERROR.  This is bad because the
  original exception doesn't show up in the log file.  Instead, just let
  exceptions escape.
- JarDirectoryStepHelper was allowing exceptions from ZipFile's
  constructor to escape.  This wouldn't be a problem, except that
  ZipFile does not include the file name in its exceptions.  Wrap it in
  a try/catch and wrap the exception if it doesn't include the file
  name.

Test Plan:
Integration test verifies that all important information shows up.  I
looked at the output manually as well and it looks good.

Reviewed By: illicitonion

fbshipit-source-id: 17d7c25