Write metadata.type with 644 permissions; don't overwrite unless needed
Summary:
Under some circumstances it's desireable to have one user able to use the build
results of another; metadata.type gets in the way because it's created with 0600
perms by default, and is always written even if it doesn't need to be.  This
diff fixes both of these issues.

Test Plan: Simulate this with a new integration test

Reviewed By: asp2insp

fbshipit-source-id: 7251b82