Added property to rules to add files that won't be built to the Xcode project
Summary:
In case we have files that are generated using a manual script that generate files for Xcode,
it's still useful to have the source files around to be able to edit them in Xcode.

This diff adds a property "extra_xcode_files" to ios_library.

```
cxx_library(
    name = "foobar",
    srcs = [...],
    ...
    extra_xcode_files = [
        "foobar.json",
    ]
)
```

Test Plan:
- unit tests
- add extra_xcode_files to an existing library and generate a project, then see the referenced files here.

Reviewed By: yiding

fbshipit-source-id: 34bbd1f