engine: fix bug causing builds to cache using stale manifests
Summary:
Since some refactorings, the build engine no longer eagerly removed existing
on-disk manifest files when performing the cache fetch.  This exposed an issue
where at the end of processing the rule, when updating the manifest and pushing
it to the cache, the build engine would use an existing, stale manifest written
by an unrelated build.  This would cause invalid entries from the stale
manifest to get written into the manifest pushed to the cache for the current
rule, and therefore cause cache poisoning.

This diff tries to address the issue by storing the manifest rule key inside the
manifest, so that we can detect when an existing manifest is stale or not.  If
the build engine sees a stale manifest on upload, it clears it and uses a new
one.

Test Plan: added unittest

Reviewed By: bertmaher

fbshipit-source-id: 317cce8