Move daemon lifecycle management into its own class, and clean up calls to getDaemon
Summary:
Daemon lifecycle management is fairly separate from the rest of main logic, so
I moved it away.

I also noticed that we call getDaemon repeatedly while initializing things in
Main, so I changed it so that it's only called once.

Also related is that we have an isDaemon boolean, when this can be informed
from an Optional<Daemon> instead. This cleans up a lot of
if-daemon-then-get-daemon-else-empty methods.

Test Plan: CI

Reviewed By: marcinkosiba

fbshipit-source-id: 4a0b792